moved no-funds to common file
