--- packages/augur-ui/src/modules/create-market/components/no-funds-error.styles.less
@@ -1,9 +1,0 @@
-@import (reference) '~assets/styles/shared';
-
-.HasError {
-  margin-top: @size-32;
-
-  > section {
-    margin-top: @size-12;
-  }
-}
